Korean BERT to Korean SBERT
https://pythonawesome.com/utilize-korean-bert-model-in-sentence-transformers-library/

There is a python package which kinda converts usual bert model to sentence bert model. In that library, it returns modules which can be loaded in SentenceTransformer package for SentenceTrasformer model(SBERT). Because of few korean SBERT exist, I tried to use usual bert model as SBERT. When you use modules as an input(tp = TopicModelDataPreparation(modules=[word_embedding_model, pooling_model])), it loads the model(usual bert model) as an SBERT model.

Warning, The warning message when no model is specified is removed. Can be improved in the future.